+++ modules/client/test/integration/tokenVoting-client/estimation.test.ts
@@ -1,8 +1,5 @@
-import * as mockedGraphqlRequest from "../../mocks/graphql-request";
-// @ts-ignore
-declare const describe, it, expect, beforeAll, afterAll;
-
 // mocks need to be at the top of the imports
+import * as mockedGraphqlRequest from "../../mocks/graphql-request";
 import "../../mocks/aragon-sdk-ipfs";
 
 import {
@@ -18,19 +15,15 @@ import {
   contextParamsLocalChain,
   SUBGRAPH_PLUGIN_INSTALLATION,
 } from "../constants";
-import * as ganacheSetup from "../../helpers/ganache-setup";
 import * as deployContracts from "../../helpers/deployContracts";
-import { Server } from "ganache";
 import { Context } from "@aragon/sdk-client-common";
 import { buildTokenVotingDAO } from "../../helpers/build-daos";
 import { createTokenVotingPluginBuild } from "../../helpers/create-plugin-build";
 
 describe("Token Voting Client", () => {
   describe("Estimation Module", () => {
-    let server: Server;
     let deployment: deployContracts.Deployment;
     beforeAll(async () => {
-      server = await ganacheSetup.start();
       deployment = await deployContracts.deploy();
       contextParamsLocalChain.daoFactoryAddress = deployment.daoFactory.address;
       contextParamsLocalChain.tokenVotingRepoAddress =
@@ -39,10 +32,6 @@ describe("Token Voting Client", () => {
         deployment.pluginSetupProcessor.address;
       contextParamsLocalChain.ensRegistryAddress =
         deployment.ensRegistry.address;
-    });
-
-    afterAll(async () => {
-      await server.close();
     });
 
     it("Should estimate the gas fees for creating a new proposal", async () => {